% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragula.R
\name{dragulaOutput}
\alias{dragulaOutput}
\alias{renderDragula}
\title{Shiny bindings for dygraph}
\usage{
dragulaOutput(outputId)

renderDragula(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from.}

\item{expr}{An expression that generates a dragula object.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\value{
RETURN_DESCRIPTION
}
\description{
Output and render functions for using dygraph within Shiny
applications and interactive Rmd documents.
}
\details{
Dragula widget.

Create dragula widget.
}
\examples{

# ADD_EXAMPLES_HERE

}
