% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragula.R
\name{useDragulajs}
\alias{useDragulajs}
\title{Register dragulaR's js functions for refreshing dragula object.}
\usage{
useDragulajs()
}
\value{
A Shiny tag list that registers the JavaScript extension.
}
\description{
This function enables the \code{js$refreshDragulaR()} JavaScript function
in Shiny, which should be called after dynamically adding elements to
a dragula container.
}
\examples{

\dontrun{
# See example for more details
library(dragulaR)
runApp(
  system.file("apps/example05-dragula-dynamic-elements", package = "dragulaR"),
  display.mode = "showcase")
}

}
