% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragula.R
\name{dragula}
\alias{dragula}
\title{Register containers to dragula.}
\usage{
dragula(x, id = NULL)
}
\arguments{
\item{x}{vector of containers ids. Their's elements will become draggable.}

\item{id}{input id to read from in shiny.}
}
\value{
Return htmlWidget. Should be used only inside shiny ui.
}
\description{
Create dragula instace to allow moving around elements of the registered containers.
}
\examples{

if(interactive()) {
  path <- system.file("apps/example01-dragula", package = "dragulaR")
  runApp(path, display.mode = "showcase")
}

}
