% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragula.R
\name{dragulaValue}
\alias{dragulaValue}
\title{Format dragula input object.}
\usage{
dragulaValue(x)
}
\arguments{
\item{x}{dragula input.}
}
\description{
This is a utility function for better formatting dragula's input.
}
\examples{

\dontrun{
# Example call:
dragulaValue(input$dragula)
}

}
