% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragula.R
\name{dragulaOutput}
\alias{dragulaOutput}
\alias{renderDragula}
\title{Dragula widget.}
\usage{
dragulaOutput(outputId)

renderDragula(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from.}

\item{expr}{An expression that generates a dragula object.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\value{
RETURN_DESCRIPTION
}
\description{
Create dragula widget.
}
\examples{

if(interactive()) {
  path <- system.file("apps/example02-input", package = "dragulaR")
  runApp(path, display.mode = "showcase")
}

}
