% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragula.R
\name{useDragulajs}
\alias{useDragulajs}
\title{Register dragulaR's js functions for refreshing dragula object.}
\usage{
useDragulajs()
}
\description{
Register dragulaR's js functions for refreshing dragula object.
}
\examples{

\dontrun{
# See example for more details
library(dragulaR)
runApp(
  system.file("apps/example05-dragula-dynamic-elements", package = "dragulaR"),
  display.mode = "showcase")
}

}
